Separated main.js into modules.
