+++ index.html
@@ -11,7 +11,10 @@
     <link rel="preconnect" href="https://api.fonts.coollabs.io" crossorigin />
     <link rel="shortcut icon" href="./assets/icons/favicon.svg" type="image/x-icon" id="favicon">
     <link href="https://api.fonts.coollabs.io/css2?family=Montserrat&display=swap" rel="stylesheet">
-    <script src="./main.js"></script>
+    <script type="module" src="./main.js"></script>
+    <script type="module" src="./modules/code_comparison.mjs"></script>
+    <script type="module" src="./modules/update_button.mjs"></script>
+    <script type="module" src="./modules/updated_since.mjs"></script>
 </head>
 <body>
     <div id="menubar">
